Move lua header parser to bin/format ##bin
